+++ app/Http/Controllers/Admin/ProductController.php
@@ -25,6 +25,7 @@ class ProductController extends Controller
         
     }
     public function stroe(Request $request){
+        if(Auth::user()->role == 'main'){
             $validatedData = $request->validate([
                 'name' => 'required|max:255',
                 'price' => 'required',
@@ -52,56 +53,63 @@ class ProductController extends Controller
             if($product->save()){
                 return response()->json(['msg' => 'Added']);
             }
-        
+            return response()->json(['msg' => 'Error adding']);
+        }
+        return response()->json(['msg'=> 'unauthorized action'], 401);
     }
 
     public function destroy($id){
-        
+        if(Auth::user()->role == 'main'){
             $product = Product::findOrFail($id);
-            $imagePath = $product->image;
+            //$imagePath = $product->image;
             if($product->delete()){
                 // delete image from folder
-                if(File::exists($imagePath)) {
-                    File::delete($imagePath);
-                }
+                // if(File::exists($imagePath)) {
+                //     File::delete($imagePath);
+                // }
                 return response()->json(['msg' => 'Deleted']);
             }
             return response()->json(['msg' => 'Error Deleting']);
-        
+        }
+        return response()->json(['msg'=> 'unauthorized action'], 401);
     }
 
     public function update(Request $request){
-        $product = Product::findOrFail($request->id);
-        $product->name = $request->name;
+        if(Auth::user()->role == 'main'){
+            $product = Product::findOrFail($request->id);
+            $product->name = $request->name;
 
-        if($request->hasFile('image'))
-        {
-            $image_name = time().'_'.rand(999,9999).''.rand(999,99999).''.rand(999,99999).'.'.$request->image->getClientOriginalExtension();
-            $request->image->move('images/products', $image_name);
-            $product->image = 'images/products/'.$image_name;
+            if($request->hasFile('image'))
+            {
+                $image_name = time().'_'.rand(999,9999).''.rand(999,99999).''.rand(999,99999).'.'.$request->image->getClientOriginalExtension();
+                $request->image->move('images/products', $image_name);
+                $product->image = 'images/products/'.$image_name;
 
-        }else{ 
-            $product->image = $request->current_image;
-            if(empty($request->current_image)){
-                $product->image = 'images/products/default.jpg';
-            }  
+            }else{ 
+                $product->image = $request->current_image;
+                if(empty($request->current_image)){
+                    $product->image = 'images/products/default.jpg';
+                }  
+            }
+
+            $product->des = $request->des;
+            $product->price = $request->price;
+            $product->category_id = $request->category_id;
+
+            if($product->save()){
+                return response()->json(['msg' => 'Updated']);
+            }
         }
-
-        $product->des = $request->des;
-        $product->price = $request->price;
-        $product->category_id = $request->category_id;
-
-        if($product->save()){
-            return response()->json(['msg' => 'Updated']);
-        }
+        return response()->json(['msg'=> 'unauthorized action'], 401);
     }
 
     public function updateStatus($status, $id){
-        
+        if(Auth::user()->role == 'main'){
             $product = Product::findOrFail($id);
             $product->status = $status;
             $product->save();
             return new ProductResource($product);
-        
+        }
+        return response()->json(['msg'=> 'unauthorized action'], 401);
     }
 }
